+++ account/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import AbstractBaseUser, BaseUserManager
 from easysurfClubs.models import Club
-from easysurfHome.models import ResidentChecklist
 
 
 class MainAccountManager(BaseUserManager):
@@ -82,3 +81,12 @@ class Account(AbstractBaseUser):
 
     def has_module_perms(self, app_label):
         return True
+
+class ResidentChecklist(models.Model):
+    resident = models.ForeignKey(Account, on_delete=models.CASCADE)
+
+    confirmed_personal_info = models.BooleanField(default=False)
+    orientation = models.BooleanField(default=False)
+    completed_survey = models.BooleanField(default=False)
+    joined_club = models.BooleanField(default=False)
+    voted_issue = models.BooleanField(default=False)
