+++ easysurfHome/models.py
@@ -10,15 +10,6 @@ class Visitor(models.Model):
 
     resident = models.ForeignKey(Account, on_delete=models.CASCADE)
 
-class ResidentChecklist(models.Model):
-    resident = models.ForeignKey(Account, on_delete=models.CASCADE)
-
-    confirmed_personal_info = models.BooleanField(default=False)
-    orientation = models.BooleanField(default=False)
-    completed_survey = models.BooleanField(default=False)
-    joined_club = models.BooleanField(default=False)
-    voted_issue = models.BooleanField(default=False)
-
 class OrientationResidentDate(models.Model):
     resident = models.ForeignKey(Account, on_delete=models.CASCADE)
     orientation_date = models.DateTimeField(default=None, blank=True, null=True)
